+++ experimental/packages/opentelemetry-exporter-metrics-otlp-grpc/test/metricsHelper.ts
@@ -19,10 +19,19 @@ import { Resource } from '@opentelemetry/resources';
 import * as assert from 'assert';
 import * as grpc from '@grpc/grpc-js';
 import { VERSION } from '@opentelemetry/core';
-import { ExplicitBucketHistogramAggregation, MeterProvider, MetricReader } from '@opentelemetry/sdk-metrics-base';
+import {
+  AggregationTemporality,
+  ExplicitBucketHistogramAggregation,
+  MeterProvider,
+  MetricReader,
+} from '@opentelemetry/sdk-metrics-base';
 import { IKeyValue, IMetric, IResource } from '@opentelemetry/otlp-transformer';
 
-export class TestMetricReader extends MetricReader {
+class TestMetricReader extends MetricReader {
+  selectAggregationTemporality() {
+    return AggregationTemporality.CUMULATIVE;
+  }
+
   protected onForceFlush(): Promise<void> {
     return Promise.resolve(undefined);
   }