+++ experimental/packages/opentelemetry-sdk-metrics-base/src/export/MetricData.ts
@@ -16,16 +16,18 @@
 
 import { HrTime } from '@opentelemetry/api';
 import { MetricAttributes } from '@opentelemetry/api-metrics';
-import { InstrumentationLibrary } from '@opentelemetry/core';
+import { InstrumentationScope } from '@opentelemetry/core';
 import { Resource } from '@opentelemetry/resources';
 import { InstrumentDescriptor } from '../InstrumentDescriptor';
 import { Histogram } from '../aggregator/types';
+import { AggregationTemporality } from './AggregationTemporality';
 
 /**
  * Basic metric data fields.
  */
 export interface BaseMetricData {
   readonly descriptor: InstrumentDescriptor;
+  readonly aggregationTemporality: AggregationTemporality;
   /**
    * DataPointType of the metric instrument.
    */
@@ -54,15 +56,15 @@ export interface HistogramMetricData extends BaseMetricData {
  */
 export type MetricData = SingularMetricData | HistogramMetricData;
 
-export interface InstrumentationLibraryMetrics {
-  instrumentationLibrary: InstrumentationLibrary;
+export interface ScopeMetrics {
+  scope: InstrumentationScope;
   metrics: MetricData[];
   errors: unknown[];
 }
 
 export interface ResourceMetrics {
   resource: Resource;
-  instrumentationLibraryMetrics: InstrumentationLibraryMetrics[];
+  scopeMetrics: ScopeMetrics[];
 }
 
 /**