+++ experimental/packages/opentelemetry-sdk-metrics-base/src/state/MeterSharedState.ts
@@ -16,9 +16,9 @@
 
 import { HrTime } from '@opentelemetry/api';
 import * as metrics from '@opentelemetry/api-metrics';
-import { InstrumentationLibrary } from '@opentelemetry/core';
-import { MetricCollectOptions } from '..';
-import { InstrumentationLibraryMetrics } from '../export/MetricData';
+import { InstrumentationScope } from '@opentelemetry/core';
+import { MetricCollectOptions } from '../export/MetricProducer';
+import { ScopeMetrics } from '../export/MetricData';
 import { createInstrumentDescriptorWithView, InstrumentDescriptor } from '../InstrumentDescriptor';
 import { Meter } from '../Meter';
 import { isNotNullish } from '../utils';
@@ -38,12 +38,12 @@ export class MeterSharedState {
   private _observableRegistry = new ObservableRegistry();
   meter: Meter;
 
-  constructor(private _meterProviderSharedState: MeterProviderSharedState, private _instrumentationLibrary: InstrumentationLibrary) {
+  constructor(private _meterProviderSharedState: MeterProviderSharedState, private _instrumentationScope: InstrumentationScope) {
     this.meter = new Meter(this);
   }
 
   registerMetricStorage(descriptor: InstrumentDescriptor) {
-    const views = this._meterProviderSharedState.viewRegistry.findViews(descriptor, this._instrumentationLibrary);
+    const views = this._meterProviderSharedState.viewRegistry.findViews(descriptor, this._instrumentationScope);
     const storages = views
       .map(view => {
         const viewDescriptor = createInstrumentDescriptorWithView(view, descriptor);
@@ -59,7 +59,7 @@ export class MeterSharedState {
   }
 
   registerAsyncMetricStorage(descriptor: InstrumentDescriptor, callback: metrics.ObservableCallback) {
-    const views = this._meterProviderSharedState.viewRegistry.findViews(descriptor, this._instrumentationLibrary);
+    const views = this._meterProviderSharedState.viewRegistry.findViews(descriptor, this._instrumentationScope);
     views.forEach(view => {
       const viewDescriptor = createInstrumentDescriptorWithView(view, descriptor);
       const aggregator = view.aggregation.createAggregator(viewDescriptor);
@@ -75,9 +75,9 @@ export class MeterSharedState {
   /**
    * @param collector opaque handle of {@link MetricCollector} which initiated the collection.
    * @param collectionTime the HrTime at which the collection was initiated.
-   * @returns the list of {@link MetricData} collected.
+   * @returns the list of metric data collected.
    */
-  async collect(collector: MetricCollectorHandle, collectionTime: HrTime, options?: MetricCollectOptions): Promise<InstrumentationLibraryMetrics> {
+  async collect(collector: MetricCollectorHandle, collectionTime: HrTime, options?: MetricCollectOptions): Promise<ScopeMetrics> {
     /**
      * 1. Call all observable callbacks first.
      * 2. Collect metric result for the collector.
@@ -94,7 +94,7 @@ export class MeterSharedState {
       .filter(isNotNullish);
 
     return {
-      instrumentationLibrary: this._instrumentationLibrary,
+      scope: this._instrumentationScope,
       metrics: metricDataList.filter(isNotNullish),
       errors,
     };